+++ src/service/stringConverter/url.ts
@@ -1,5 +1,5 @@
 import { TokenInfo } from "../codeParser";
-import { isStringToken } from "../literalParser/interface";
+import { isStringToken, isUnknownToken } from "../literalParser/interface";
 import { StringConverter, StringConverterConvertResult, StringConverterMatchResult, StringConverterMeta, StringConverterOptions } from "./interface";
 
 interface URLMatchByProduct {
@@ -26,9 +26,9 @@ export class UrlParser implements StringConverter<URLMatchByProduct> {
     };
 
     match(tokenInfo: TokenInfo, options?: StringConverterOptions): StringConverterMatchResult<URLMatchByProduct> {
-        // if (!isStringToken(tokenInfo.type)) {
-        //     return { matched: false };
-        // }
+        if (!isStringToken(tokenInfo.type) && !isUnknownToken(tokenInfo.type)) {
+            return { matched: false };
+        }
 
         const parseURL = (str: string) => {
             try {